% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_string.R
\name{cleanString}
\alias{cleanString}
\title{Clean News Article}
\usage{
cleanString(text)
}
\arguments{
\item{text}{The string to be cleaned. Either character or character vector.}
}
\value{
The cleaned and transformed text string or strings vector, depending on the input format.
}
\description{
`cleanString` cleans scrapped news articles text strings by removing and/or replacing special characters and punctuation. It also trims the string and encodes it with "UTF-8".
}
\examples{
data(news)
cleanString(news$text)
}
